requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
adbutils=0.16.2=pypi_0
apkutils2=1.0.0=pypi_0
asn1crypto=1.5.1=pypi_0
binutils_impl_linux-64=2.35.1=h27ae35d_9
binutils_linux-64=2.35.1=h454624a_30
blas=1.0=mkl
blosc=1.21.0=h8c45485_0
brotli=1.0.9=he6710b0_2
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.4.26=h06a4308_0
cached-property=1.5.2=pypi_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
cfitsio=3.470=hf0d0db6_6
charls=2.2.0=h2531618_0
charset-normalizer=2.0.12=pypi_0
cigam=0.0.3=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cprint=1.2.2=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2022.2.1=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
deprecated=1.2.13=pypi_0
deprecation=2.1.0=pypi_0
docopt=0.6.2=pypi_0
expat=2.4.4=h295c915_0
filelock=3.7.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2022.2.0=pyhd3eb1b0_0
gcc_impl_linux-64=9.3.0=h6df7d76_17
gcc_linux-64=9.3.0=h1ee779e_30
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gxx_impl_linux-64=9.3.0=hbdd7822_17
gxx_linux-64=9.3.0=h7e70986_30
icu=58.2=he6710b0_3
idna=3.3=pypi_0
imagecodecs=2021.8.26=py39h4cda21f_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2019.4=243
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonlines=1.2.0=pypi_0
jxrlib=1.1=h7b6447c_2
kernel-headers_linux-64=2.6.32=he073ed8_15
kiwisolver=1.3.2=py39h295c915_0
krb5=1.19.2=hac12032_0
lark=1.1.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libcurl=7.82.0=h0b77cf5_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-devel_linux-64=9.3.0=hb95220a_17
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libmklml=2019.0.5=h06a4308_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-devel_linux-64=9.3.0=hf0c5c8d_17
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h74e7548_1
libzopfli=1.0.3=he6710b0_0
locket=0.2.1=py39h06a4308_2
logzero=1.7.0=pypi_0
lxml=4.8.0=pypi_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
msbase=0.3.4=pypi_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
networkx=2.7.1=pyhd3eb1b0_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.22.3=pypi_0
numpy-base=1.19.2=py39h2ae0177_0
opencv-contrib-python=4.5.5.64=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1o=h7f8727e_0
packaging=20.9=pypi_0
partd=1.2.0=pyhd3eb1b0_1
path=16.4.0=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
pipreqs=0.4.11=pypi_0
progress=1.6=pypi_0
ptyprocess=0.7.0=pypi_0
py=1.11.0=pypi_0
pyaxmlparser=0.3.27=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyelftools=0.28=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.7.1=cpu_py39h6a09485_0
pywavelets=1.3.0=py39h7f8727e_0
pyyaml=5.3.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
retry=0.9.2=pypi_0
scikit-image=0.19.2=py39h51133e4_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.6.2=py39h91f5cce_0
setuptools=61.2.0=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.38.3=hc218d9a_0
sysroot_linux-64=2.12=he073ed8_15
tabulate=0.8.9=pypi_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torch=1.12.0.dev20220509+cu116=pypi_0
torchvision=0.8.2=cpu_py39ha229d99_0
tornado=6.1=py39h27cfd23_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
uiautomator2=2.16.14=pypi_0
urllib3=1.26.9=pypi_0
utils=1.0.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
whichcraft=0.6.1=pypi_0
wrapt=1.14.1=pypi_0
xmltodict=0.13.0=pypi_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
yarg=0.1.9=pypi_0
zfp=0.5.5=h295c915_6
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
